abstract common module
